Extract Naive TreeSHAP implementation into a separate class
